// apis/api/src/models/RoleName.ts
/**
 * Names of all the available roles in the system
 */
export enum RoleName {
  SuperAdmin = 'Super Admin',
  CompanyAdmin = 'Company Admin',
  BasicUser = 'Basic User',
  OrderManager = 'Order Manager',
  SalesManager = 'Sales Manager',
  Agronomist = 'Agronomist',
  SalesRep = 'Sales Rep',
  Applicator = 'Applicator',
  Blender = 'Blender',
  Dispatcher = 'Dispatcher',
  Driver = 'Driver',
  Puller = 'Puller', // Soil Sample Puller
  Contact = 'Contact',
}